DDF-1652 Refactored itest project name and moved test classes into new packages.
