+++ distribution/test/itests/test-itests-ddf/pom.xml
@@ -21,8 +21,8 @@
         <artifactId>itests</artifactId>
         <version>2.9.0-SNAPSHOT</version>
     </parent>
-    <artifactId>test-itests-catalog</artifactId>
-    <name>DDF :: Test :: Integration :: Catalog</name>
+    <artifactId>test-itests-ddf</artifactId>
+    <name>DDF :: Integration  :: Test</name>
 
     <dependencies>
         <dependency>
