+++ distribution/test/itests/test-itests-ddf/src/test/java/ddf/test/itests/common/Library.java
@@ -12,7 +12,7 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-package ddf.catalog.test;
+package ddf.test.itests.common;
 
 import org.apache.commons.lang.StringUtils;
