+++ distribution/test/itests/test-itests-ddf/src/test/java/ddf/test/itests/common/SecurityPolicyConfigurator.java
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  **/
-package ddf.catalog.test;
+package ddf.test.itests.common;
 
 import java.util.Map;
 import java.util.concurrent.Callable;
