+++ distribution/test/itests/test-itests-ddf/src/test/java/ddf/test/itests/platform/TestXACML.java
@@ -11,7 +11,7 @@
  * is distributed along with this program and can be found at
  * <http://www.gnu.org/licenses/lgpl.html>.
  */
-package ddf.catalog.test;
+package ddf.test.itests.platform;
 
 import java.io.IOException;
 import java.util.Dictionary;
