+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.catalog.solr.external.SolrHttpCatalogProvider.config
