+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.security.sts.client.configuration.config
