+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.test.itests.platform.TestPlatform.invalid.config
