+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.test.itests.platform.TestPlatform.msf.1.config
@@ -1,4 +1,5 @@
-id="ddf.catalog.test.TestPlatform.nopid"
+service.factoryPid="ddf.test.itests.platform.TestPlatform.msf"
+id="ddf.test.itests.platform.TestPlatform.msf.1"
 
 property.string="string"
 property.boolean.true=B"true"
