+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.test.itests.platform.TestPlatform.msf.2.config
@@ -1,5 +1,5 @@
-service.pid="ddf.catalog.test.TestPlatform.new"
-id="ddf.catalog.test.TestPlatform.new"
+service.factoryPid="ddf.test.itests.platform.TestPlatform.msf"
+id="ddf.test.itests.platform.TestPlatform.msf.2"
 
 property.string="string"
 property.boolean.true=B"true"
