+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.test.itests.platform.TestPlatform.new.config
@@ -1,5 +1,5 @@
-service.pid="ddf.catalog.test.TestPlatform.startup"
-id="ddf.catalog.test.TestPlatform.startup"
+service.pid="ddf.test.itests.platform.TestPlatform.new"
+id="ddf.test.itests.platform.TestPlatform.new"
 
 property.string="string"
 property.boolean.true=B"true"
