+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.test.itests.platform.TestPlatform.nopid.config
@@ -1,5 +1,4 @@
-service.factoryPid="ddf.catalog.test.TestPlatform.msf"
-id="ddf.catalog.test.TestPlatform.msf.2"
+id="ddf.test.itests.platform.TestPlatform.nopid"
 
 property.string="string"
 property.boolean.true=B"true"
