+++ distribution/test/itests/test-itests-ddf/src/test/resources/ddf.test.itests.platform.TestPlatform.startup.config
@@ -1,5 +1,5 @@
-service.factoryPid="ddf.catalog.test.TestPlatform.msf"
-id="ddf.catalog.test.TestPlatform.msf.1"
+service.pid="ddf.test.itests.platform.TestPlatform.startup"
+id="ddf.test.itests.platform.TestPlatform.startup"
 
 property.string="string"
 property.boolean.true=B"true"
