+++ distribution/test/itests/test-itests-ddf/src/test/resources/metacard1.xml
