+++ distribution/test/itests/test-itests-ddf/src/test/resources/metacard2.xml
